Move all deps to devDependencies
It seems that electron-builder will bundle everything that's on the
dependencies list into the compiled app. Because we use webpack to
compile everything, this is just copying unecessary code to the app,
increasing the bundle size. Moving everything from to the
devDependencies reduced the app's size from 240MB to 180MB.