Improve selection color
I can't customize the color of a selection's text, only the background.
This makes it hard to use one of the three user colors, because they may
require a different text color. Instead, I'm transforming the color to
HSL and adjusting the lightness. This will give it just enough contrast
while not looking out of place.