Use SQLite
I'm replacing lowdb and elasticlunr with SQLite to have full-text
search. Every tool I tried doesn't come close to it, and I've tried a
few.

Unfortunately, I couldn't find a small tool to handle migrations and
build queries that worked with electron. It could have been because of
my lack of experience with these tools, but why does every tool that has
migrations need a migration folder? It's super hard to work with folders
when you have typescript, webpack and electron-builder. Anyway, it
wasn't much code, so I just wrote my own migration code and my queries
in SQL. It's quite refreshing to not have a layer of
abstraction/indirection between my queries and the database.
I'm replacing lowdb and elasticlunr with SQLite to have full-text search. Every tool I tried doesn't come close to it, and I've tried a few.

Because of this change, I had to remove the sync functionality. I could
rewrite it, but I thought that maybe we can see if anyone asks for it.