Only clear the editor after a success message
Something weird is happening, and people are losing files. I'm not sure,
but I'm confident that it's related to this. When a note is saved, the
editor is cleared, assuming that the save was successful. But there can
be errors writing to the database, preventing the note from being saved,
and therefore it's lost forever. This change should prevent this because
it only clears the buffer if the save was successfull.